requirements.txt
wearipedia
psycopg2-binary
python-dotenv
